chore: update lib to be interface
Update libs folder types to be interfaces isntead of classes.
This allows Angualr to not depend on type-graphql
which uses fs under the hood. As fs is only a node module
Angualr cannot resolve the dependencies and the build ends up breaking.
From here on all shared types will be interfaces and Nest types
will implement the interface to ensure no tyeps are being overwritten.

Related to #333 and #310