chore(cleanup): set typescript compiler to strict mode
Sets the TSC compiler to strict mode to ensure better coverage of the
application while in development. Adds several lines to ConfigModule
and DatabaseModule setup files to allow for type assertion. Adds type
assertions for most if not all DTOs and classes to ensure that TSC does
not complain. Removed references of class[key] and uses spread operator
in place instead. Add tests for all graphql-return files.