feat(redis): creates start of redis service
The redis service now works with Observables thanks to
[this issue](https://github.com/ReactiveX/rxjs/issues/1720)
and allows for easy integration into the rest of the server.
Currently, the methods are memoized, but they may need to be
deferred instead depending on how the memoization affects
the output. SessionIds and UserIds are properly being stored
in the case of Google Logins. Now to move back to the Local
login strategy. More than likely, a "LocalModule" will be
created under the "auth" folder, similar to the "GoogleModule"
which will allow the logic to be encapsulated. A
"CookieModule" or similar should probably also be created to
remove duplication off the cookie setting code, including
the call to Redis to make everything smoother in the long run.