feat(auth): removed all uses of passport code
Now with using the custom authentication code, there's no more
concern about when passport sets things like req.user or
needing to call req.isAuthenticated. Some functionalities
still need to be replcaed, like the guards and the session
check, but on a whole the custom authentication is making great
progress.