feat(refresh): session refresh now implemented
Two cookies are now sent back to the client on login/signup
a session id and a refresh cookie. The ID is available for an
hour and the refresh is good for 14 days. After the session
token expires the refresh token is checked and will give a new
session and refresh token. It should probably be implemented
that only the session id token is updated, but that's an
enhancement for later.