feat(database): adds serialization to query directly
The `DatabaseSerivice` now takes in a type to know what to serialize
the result to. This is essentially a primitive ORM that I'm in control
of, which is why I'm happy with it. It also still makes use of raw SQL

This commit also adds the serialization to the `AuthService` and its
methods